refactor(kernel/dispatch): move ObserverRegistry + NativeHookRegistry into core/dispatch/
Aligns the dispatch primitive with the §4 ownership model in
KERNEL-ARCHITECTURE.md: every other primitive (vfs_router, lock,
file_watch, …) keeps its full state under core/{primitive}/. Dispatch
was the only one whose registries lived in kernel/mod.rs.

  - KernelObserverEntry / KernelObserverRegistry → ObserverEntry /
    ObserverRegistry (the Kernel prefix was disambiguation against
    PyKernel etc., unnecessary inside the dispatch module).
  - NativeHookEntry / NativeHookRegistry kept their names.
  - Kernel.observers field type: Mutex<KernelObserverRegistry> →
    Mutex<ObserverRegistry>; native_hooks field type unchanged.

Visibility stays pub(crate). Behavior unchanged.
